Update edit-escrow-hints.md
I've made a couple updates that make the tests appear more like the refund test (which you recommend copying). If the format you originally listed here is needed might be good to make reference to the changes. It might be because I tried to copy the refund test or it might be other mistakes but i'm unable to run tests successfully. It seems like I'm supposed to compile before running the tests, is this so? Actually I just ran `cargo wasm`, it compiled successfully, then ran `RUST_BACKTRACE=1 cargo unit-test` it was successful but only ran the original 5 tests.
My repo with errors can be seen here: https://github.com/okwme/cosmwasm-examples